Step 8: Split compiler into files and modify Makefile
With this patch pcc.c is splitted into multiple files that contain the
specific functionalities of the compier, i.e., parser, tokenizer and
code generator. pcc.h holds the declarations of all the functions,
structs and global variables at the moment and every .c file includes
it. Makefile is also modified to support the splitted files.
