+++ Makefile
@@ -1,6 +1,11 @@
 CFLAGS=-std=c11 -g -static
+SRCS=$(wildcard *.c)
+OBJS=$(SRCS:.c=.o)
 
-pcc: pcc.c
+pcc: $(OBJS)
+	$(CC) -o pcc $(OBJS) $(LDFLAGS)
+
+$(OBJS): pcc.h
 
 test: pcc
 	./test.sh
